Qwen3-8B FV rerun: bf16 load, generate_str filters, ambiguous whole-answer correctness, sharding driver
- model_utils: load Qwen in bf16 on CUDA (Qwen3 is bf16-native; fp16 risks NaN)
- eval_utils.build_eval_item: fix generate_str filter crash (list target -> get_answer_id)
- compute_multitask_top_aie_heads: --generate_str/--metric full-answer ICL filter
- varicl_utils + compute_multitask_varicl_heads: generate_str variant of correctness filter
- compute_ambiguous_constrained_fv: greedy-generate + exact-match correctness (numeric-safe)
- run_qwen3_fv_all.py: N-GPU sharded orchestrator (waves A-E, --only/--dry_run)

+++ src/utils/model_utils.py
@@ -119,7 +119,8 @@ def load_gpt_model_and_tokenizer(model_name:str, device='cuda', revision=None):
     elif 'qwen' in model_name.lower():  # covers Qwen2 / Qwen2.5 / Qwen3 (same hook layout)
         tokenizer = AutoTokenizer.from_pretrained(model_name)
         tokenizer.pad_token = tokenizer.eos_token
-        model_dtype = torch.float16 if str(device).startswith('cuda') else torch.float32
+        # Qwen3 is bf16-native; fp16 risks NaN/inf logits on some layers. Use bf16 on CUDA.
+        model_dtype = torch.bfloat16 if str(device).startswith('cuda') else torch.float32
         if revision is not None:
             model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=model_dtype, revision=revision).to(device)
         else:
